Moved the MesssageProcessor.h file into its parent directory and changed the pointer to the communicator in the MessageProcessor class to a shared_ptr
