+++ include/communication/Communicator.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "SSLSocket.h"
-#include "messageProcessor/MessageProcessor.h"
+#include "MessageProcessor.h"
 #include "message/Message.h"
 #include <memory>
