+++ include/worker/WorkerMP.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <communication/messageProcessor/MessageProcessor.h>
+#include <communication/MessageProcessor.h>
 
 class WorkerMP : public MessageProcessor {
 public:
